pyproject.toml
[project]
name = "zinc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.2",
    "django==4.1.5",
    "django-reset-migrations==0.4.0",
    "djangorestframework>=3.15.1",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "python-json-logger==3.2.1",
    "ujson>=5.10.0",
]
